% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citbi.R
\docType{data}
\name{citbi}
\alias{citbi}
\title{Title: Raw citbi data set}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 30379 rows and 26 columns.
}
\source{
STAT 133
}
\usage{
citbi
}
\description{
Dataset on brain trauma patients
}
\keyword{datasets}
